+++ app/src/main/java/com/test/mainli/MainActivity.java
@@ -8,15 +8,15 @@ import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentPagerAdapter;
 import android.support.v4.view.ViewPager;
-import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.MenuItem;
 
+import com.test.mainFragment.MultiRecyclerViewFragment;
+import com.test.mainFragment.RecyclerViewFragment;
+
 import butterknife.Bind;
 import butterknife.ButterKnife;
-import mainFragment.MultiRecyclerViewFragment;
-import mainFragment.RecyclerViewFragment;
 
 public class MainActivity extends AppCompatActivity {
     @Bind(R.id.toolbar)
